+++ src/helpers/element/index.js
@@ -1,0 +1,7 @@
+import * as createElement from "./createElement.js";
+import * as htmlToElement from "./htmlToElement.js";
+
+export const Element = {
+  ...createElement,
+  ...htmlToElement
+};